Shifter FBX tool: refactoring first pass
